==> 01.2-Netflix_add_zh_name.Rmd <==
---
title: "Netflix_add_zh_name"
author: "快樂RRRRR"
date: "2022/05/19"
output:
  html_document:
    number_sections: no
    theme: united
    highlight: tango
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, results = 'markup', comment = '#>', error = TRUE)
```

### 從 TMDB 取得對應的中文劇名

```{r message=FALSE, warning=FALSE}
### library
library(tidyverse)
library(rvest) #read_html html_nodes(), html_text()

### Param & Var
Netflix_df <- read_csv("concatALL.csv") # 讀取檔案
Netflix_df <- read_csv("concatALL_more.csv") # 讀取檔案
Res <- tibble()
netflix_en_id <- Netflix_df$id %>% unique # 確認戲劇是唯一的
Netflix_df %>% View
#count = 0
netflix_en_id %>% length
### Main Program
for (i in c(1:length(netflix_en_id))) { # 建議多加 try catch 避免 http error
  movie_id = netflix_en_id[i]
  if ( identical(as.logical(movie_id), NA) == TRUE ) next
  url <- str_c("https://www.themoviedb.org/tv/", movie_id, "?language=zh-TW")
  movie_zh_name <- url %>% read_html() %>% html_nodes("section > div.title.ott_true > h2 > a") %>% html_text()
  T <- tibble(movie_id, movie_zh_name)
  Res <- bind_rows(Res, T)
  random_num <- runif(1, 3, 8)
  Sys.sleep(random_num)
  print(paste(i , netflix_en_id %>% length))
  #print(Res)
}
Res <- Res %>% mutate( id = movie_id ) %>% select(id, movie_zh_name)
Netflix_df_zh <- Netflix_df %>% left_join(Res)
Netflix_df_zh %>% 
  select(Year, Week, Ranking, name, movie_zh_name, everything()) %>%
  write_csv("Netflix_df_zh-name_more.csv")

# 87,101(?)可能有問題
# 有些跑NA看我們能不能手動補，抓資料的時候不知道有少數三四個會跑error，就有一些資訊斷掉QQQQ
```
```{r}
Netflix_df_zh_all <- read_csv("Netflix_df_zh-name_v0528.csv")
t %>% tail
t %>% mutate(tag = str_c(Year, "-",  Week, "-", Ranking)) %>%
  select(tag, everything()) 
```



# 韓國的資料
```{r}
tv_korea_NA_info <- read_csv("bak/手動處理KR_done.csv")
tv_korea_NA_info #(movie_zh_name)



netflix_en_id <- tv_korea_NA_info$id %>% unique # 確認戲劇是唯一的
#count = 0
netflix_en_id %>% length
### Main Program
Res <- tibble()
for (i in c(1:length(netflix_en_id))) { # 建議多加 try catch 避免 http error
  movie_id = netflix_en_id[i]
  if ( identical(as.logical(movie_id), NA) == TRUE ) next
  url <- str_c("https://www.themoviedb.org/tv/", movie_id, "?language=zh-TW")
  movie_zh_name <- url %>% read_html() %>% html_nodes("section > div.title.ott_true > h2 > a") %>% html_text()
  if ( identical(movie_zh_name, character(0)) ) movie_zh_name <- url %>% read_html() %>% html_nodes("section > div.title.ott_false > h2 > a") %>% html_text() # 發現有些抓不到的原因是class錯了，因此再設定一個if嘗試抓抓看
  T <- tibble(movie_id, movie_zh_name)
  Res <- bind_rows(Res, T)
  random_num <- runif(1, 3, 8)
  Sys.sleep(random_num)
  print(paste(i , netflix_en_id %>% length))
  #print(Res)
}
Res <- Res %>% rename(id=movie_id)

Res

Res$movie_zh_name <- str_replace_all(Res$movie_zh_name,"Octonauts: Above & Beyond", "海底小縱隊：登陸探險")
Res$movie_zh_name <- str_replace_all(Res$movie_zh_name,"The Raincoat Killer: Chasing a Predator in Korea", "韓國雨衣殺手：全面追緝柳永哲")
```
