==> 02-Scrap_TMDB_netflix_actor.Rmd <==
---
title: "Scrap_TMDB_netflix_actor"
author: "快樂RRRRR"
date: "2022/05/29"
output:
  html_document:
    number_sections: no
    theme: united
    highlight: tango
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, results = 'markup', comment = '#>', error = TRUE)
```

# 從 TMDB 取得對應的演員、導演

```{r message=FALSE, warning=FALSE}
### library
library(tidyverse)
library(rvest) #read_html html_nodes(), html_text()

### Param & Var

Netflix_df <- read_csv("data_storage/Netflix_Taiwan_Ranking.csv") # 讀取檔案
NetflixKR_id_zhname_df <- Netflix_df %>% select(id, movie_zh_name, origin_country) %>% distinct() %>% drop_na() %>% filter(origin_country == "KR")

length_id <- NetflixKR_id_zhname_df$id%>% length
Actor_List <- tibble()
Director_List <- tibble()
### Main Program
for (i in c(1:length_id)) { # 建議多加 try catch 避免 error
  url <- str_c("https://www.themoviedb.org/tv/", NetflixKR_id_zhname_df$id[i], "?language=zh-TW")
  print(url)
  print(paste(i, round(100* i / length_id, 3) , "%"))
  director <- url %>% read_html() %>% html_nodes("#original_header > div.header_poster_wrapper.true > section > div.header_info > ol > li:nth-child(n) > p:nth-child(1) > a") %>% html_text()
    
  actor <- url %>% read_html() %>% html_nodes("#cast_scroller > ol > li:nth-child(n) > p:nth-child(2) > a") %>% html_text() 
  Actor_List <- bind_rows(Actor_List, tibble(NetflixKR_id_zhname_df$id[i], actor))
  Director_List <- bind_rows(Director_List, tibble(NetflixKR_id_zhname_df$id[i], director))
  #break
  random_num <- runif(1, 10, 20) #隨機睡覺取值
  Sys.sleep(random_num) #睡覺中
  remove(random_num) #釋放變數
}
Actor_List <- Actor_List %>% select(id = `NetflixKR_id_zhname_df$id[i]`, actor)
Director_List <- Director_List  %>% select(id = `NetflixKR_id_zhname_df$id[i]`, director)

Netflix_KR_Actor_df <- Netflix_df %>% filter( origin_country == "KR") %>% left_join(Actor_List) %>% select(Year, Week, Ranking, TV, movie_zh_name, id, category, actor, imdb_score )


Netflix_KR_Director_df <- Netflix_df %>% filter( origin_country == "KR") %>% left_join(Director_List) %>% select(Year, Week, Ranking, TV, movie_zh_name, id, category, director, imdb_score )



```

# 導演還需手動調整NA

```{r}
Netflix_KR_Director_df %>%
  select(id, Week, Ranking, movie_zh_name,director) %>%
  count(director) %>%
  arrange(desc(n))

## NA 有點多 - 手動調整值
#Test <- Netflix_KR_Director_df %>% mutate(num = row_number()) %>%select(num, movie_zh_name, director)
#Test %>% filter(movie_zh_name == "謎霧")%>% select(num) %>% as.character() # 看到底缺的NA在哪個row

Netflix_KR_Director_df[ c(14, 19, 30),	 ]$director = "Hong Won Ki" #韓國都市怪談
Netflix_KR_Director_df[ c(63,98),        ]$director = "Kim Min-kyeong" #Do Do Sol Sol La La Sol
Netflix_KR_Director_df[ c(71, 74, 83, 87, 91, 93, 97, 100, 103),       ]$director = "Nam Gun" #私生活
Netflix_KR_Director_df[ c(102, 105, 106, 111, 114, 116, 120, 124, 128, 130, 132, 136, 143, 148),       ]$director = "Yoo Seon-dong"#驅魔麵館Park Bong-seop
Netflix_KR_Director_df[ c(108, 109, 113, 118, 121, 125),       ]$director = "Lee Eung Bok"#甜蜜家園
Netflix_KR_Director_df[ c(117, 123, 126, 129, 131, 133, 139),  ]$director = "Lee Jae-hoon"#奔向愛情
Netflix_KR_Director_df[ c(138, 142, 146, 151, 155, 161, 165, 169, 174, 175, 178, 187, 197),       ]$director = "Jin Hyuk"#薛西弗斯的神話
Netflix_KR_Director_df[ c(194, 200),     ]$director = "Choi Do-hoon"#附身
Netflix_KR_Director_df[ c(251, 264, 272, 280, 285, 296, 303, 309),       ]$director = "Jo Young Gwang"#羽球少年團
Netflix_KR_Director_df[ c(291, 302, 308, 314, 324),   ]$director = "Park Joon-hwa"#金秘書為何那樣
Netflix_KR_Director_df[ c(325, 333),                  ]$director = "Jo Soo Won"#天才醫生車耀漢
Netflix_KR_Director_df[ c(393, 400, 406, 412, 425),   ]$director = "Lee Jung-Heum"#神探具景伊
Netflix_KR_Director_df[ 395,                          ]$director = "Choi Sung-Bum"#我的ID是江南美人
Netflix_KR_Director_df[ c(397, 404, 411),             ]$director = "Mo Wan Il"#謎霧


Netflix_KR_Director_df

```

# 調整兩個df韓文的英文字

```{r}
Netflix_KR_Director_df$director <- Netflix_KR_Director_df$director %>% str_replace_all("\\W+", " ") %>% str_to_lower()
Netflix_KR_Actor_df$actor <- Netflix_KR_Actor_df$actor %>% str_replace_all("\\W+", " ") %>% str_to_lower()

write_csv(Netflix_KR_Director_df, "Netflix_KR_Director_df.csv")
write_csv(Netflix_KR_Actor_df, "Netflix_KR_Actor_df.csv")
```

# 演員/導演評比

```{r}
Netflix_KR_Actor_df <- read_csv("data_storage/Netflix_KR_Actor_df.csv")
Netflix_KR_Actor_df %>%
  select(actor, actor_number) %>%
  filter(actor_number < 5) %>%
  count(actor) %>% arrange(desc(n))
  
Netflix_KR_Actor_df %>%
  select(TV, actor, actor_number) %>%
  filter(actor_number < 5) 

Netflix_KR_Actor_df %>%
  select(TV,movie_zh_name, actor, actor_number) 

Netflix_KR_Actor_df %>%
  select(movie_zh_name, actor) %>%
  filter(actor == "Song Kang")



  
  group_by()
Netflix_KR_Director_df %>%
  select(Year, Week, Ranking, director) %>%
  count(director) %>%
  arrange(desc(n))

```
