==> 03.1-Dcard-parser.Rmd <==
---
title: "parse_netflix_dcard"
author: "快樂RRRRRR"
date: "2022/04/11"
output:
  html_document:
    number_sections: no
    theme: united
    highlight: tango
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, results = 'hold', comment = '#>', error = TRUE)
```


# Parsing Dcard API
```{r}
### Library 
library(jsonlite)
library(tidyverse)
library(stringr)
```
Dcard API 說明文件：https://blog.jiatool.com/posts/dcard_api_v2/


# 抓取文章列表
```{r}
board <- 'tvepisode' # // 填寫所選定之看板


Dcard_Netflix_Article_df <- tibble()
url_init <- 
    str_c( "https://www.dcard.tw/service/api/v2/forums/", board, "/posts?popular=false&limit=100" )
total_page <- 600
for (i in c(150:150)) {
  #print(url_r)
  if (i == 0) { # 第一輪
    Dcard_Netflix_Article_df <- url_init %>% fromJSON() %>% as_tibble()
    random_num <- runif(1, 10, 20) #隨機睡覺取值
    Sys.sleep(random_num) #睡覺中
    remove(random_num) #釋放變數
    
    print(paste(i,'RUN    ', 100*i/total_page, "%")) # 進度條
    next
  }
  last <- tail(Dcard_Netflix_Article_df, 1) %>% select(id) %>% as.character() #取得最後一篇文章的 id
  url_paging <- str_c(url_init, "&before=", last) # 加在url中
  df_tmp <- fromJSON(url_paging) %>% as_tibble()
  Dcard_Netflix_Article_df <- bind_rows(Dcard_Netflix_Article_df, df_tmp)
  
  random_num <- runif(1, 10, 20) #隨機睡覺取值
  Sys.sleep(random_num) #睡覺中
  remove(random_num) #釋放變數
  print(paste(i,'RUN    ', 100*i/total_page, "%"))
} # for (i in c(0:total_page))


Dcard_Netflix_Article_df %>% tail(5)
Dcard_Netflix_Article_df_tvepisode <- Dcard_Netflix_Article_df
saveRDS(Dcard_Netflix_Article_df_tvepisode, "Dcard_Netflix_Article_df_tvepisode.rds")

```


# 抓取文章內容 (測試)
先確認要抓哪一部劇 -> 從Dcard title 中 filter 出來 -> 再去一篇一篇抓內文(03.2)
```{r}
#Dcard_Netflix_Article_df$id %>% length
Netflix_df_zh_all <- read_csv("Netflix_df_zh-name_v0528.csv")
#Netflix_df_zh_all$movie_zh_name %>% unique %>% str_replace_all("\\W+", "")

Dcard_Netflix_Article_df_tvepisode %>%
  filter( str_detect(title, "文森佐") | str_detect(title, "黑道律師"))

Dcard_Netflix_Article_df_tvepisode %>%
  filter( str_detect(title, "海岸村"))

Dcard_Netflix_Article_df %>%
  filter( str_detect(title, "婚詞離曲"))

Dcard_Netflix_Article_df %>%
  filter( str_detect(title, "驅魔"))

Dcard_Netflix_Article_df %>%
  filter( str_detect(title, "二十五") | str_detect(title, "2521"))

Dcard_Netflix_Article_df %>%
  filter( str_detect(title, "機智醫生"))



Dcard_df_netflix <- readRDS("data_storage/Dcard/Dcard_Netflix_Article_df_netflixforum.rds")
Dcard_df_tv <- readRDS("data_storage/Dcard/Dcard_Netflix_Article_df_tvepisodeforum.rds")
Dcard_df_netflix %>%
  filter( str_detect(excerpt, "宋江"))

Dcard_df_tv %>%
  filter( str_detect(excerpt, "宋江"))
```
